CRUDS acabados, mañana hacer las paginas php

+++ app/Http/Controllers/CategoryController.php
@@ -7,19 +7,21 @@ use App\Models\Category;
 
 class CategoryController extends Controller
 {
-    public function index(){
+    public function index()
+    {
         return view('Categoria.categoria');
     }
 
-    public function createCategory(Request $request){
+    public function createCategory(Request $request)
+    {
         $date = $request->validate(
             [
-            'idUser'=>'required',
-            'name'=>'required'
+                'idUser' => 'required',
+                'name' => 'required'
             ],
             [
-                'idUser'=>'El campo idUser es obligatorio',
-                'name'=>'El campo nombre es obligatorio'
+                'idUser' => 'El campo idUser es obligatorio',
+                'name' => 'El campo nombre es obligatorio'
             ]
         );
 
@@ -28,29 +30,44 @@ class CategoryController extends Controller
         $category->name = $request->name;
         $category->save();
 
-        return response()->json(['status'=> 'success','category'=> $category]);
+        return response()->json(['status' => 'success', 'category' => $category]);
     }
 
-    public function updateCategory(Request $request){
-        $date = $request->validate([
-            'idUser'=> 'required',
-            'name'=> 'required'
-        ],
-        [
-            'idUser'=> 'Este campo es necesario',
-            'name'=> 'No puede estar vacio este campo'
-        ]
-    );
+    public function updateCategory(Request $request)
+    {
+        $date = $request->validate(
+            [
+                'idUser' => 'required',
+                'name' => 'required'
+            ],
+            [
+                'idUser' => 'Este campo es necesario',
+                'name' => 'No puede estar vacio este campo'
+            ]
+        );
 
-        $category = Category::find($request->idUser);
+        $category = Category::find($request->id);
+
+        if (!$category) {
+            return response()->json(['status' => 'error', 'message' => "Usuario no encontrado"], 404);
+        }
+
         $category->name = $request->name;
         $category->save();
-    return response()->json(['status'=> 'success','category'=> $category]);
+
+        return response()->json(['status' => 'success', 'category' => $category]);
     }
 
-    public function deleteCategory(Request $request){
-        $category = Category::find($request->idUser);
+    public function deleteCategory(Request $request)
+    {
+        $category = Category::find($request->id);
+
+        if (!$category) {
+            return response()->json(['status' => 'error', 'message' => "Usuario no encontrado"], 404);
+        }
+
         $category->delete();
-        return response()->json(['status'=> 'success','category'=> $category]);
+
+        return response()->json(['status' => 'success', 'category' => $category]);
     }
 }
